Fix problem with resolving global variables when in property definition
Because of the construction of Environments, builtin types (either
BuiltIn or builtin AggregateTypes such as quest_config) have no parent
scopes. When visitting a propertyDefinition node, the AggregateType of
the defined object will be top on stack and will be used for resolving.
Because it has no parent scope, references to global variables can not
be resolved in this context. To mittigate this, a manual stack walk is
implemented in the `resolve` method of SymbolTableParser