Add environments to store types and function definitions (#109)
* Add interface definition for environment

This allows for dynamic loading of types and functions (which improves testability and
enables the creation of dedicated environments for different use cases)

* Create standard game environment

* Use environment in SymbolTableParser

* Fix problem with resolving global variables when in property definition

Because of the construction of Environments, builtin types (either
BuiltIn or builtin AggregateTypes such as quest_config) have no parent
scopes. When visitting a propertyDefinition node, the AggregateType of
the defined object will be top on stack and will be used for resolving.
Because it has no parent scope, references to global variables can not
be resolved in this context. To mittigate this, a manual stack walk is
implemented in the `resolve` method of SymbolTableParser

* Include setup of symbolTableParser in Helpers for tests

* Include setup of symbolTableParser in getQuestConfig of DSLInterpreter

* Add helper method to get symtable with loaded types

* Fix: IScope does not define resolve with resolveInParent option

* spotless apply

Co-authored-by: mr <malt.r@mail.de>