Produs AdvancedDungeon. add Array Level (#1971)
This draft is an overview of the level design for the java/advanced
dungeon.

!!!Not ready to be merged!!!

The created levels are titled "advanced_<...>.level. A first
implementation of the monster and component spawn is realized in the
class "arraylevel". Hotload functionality of the code will be realized
in the following days.

@AMatutat 
![CleanShot 2025-05-09 at 18 26
02@2x](https://github.com/user-attachments/assets/40c89eee-b245-4cd7-9f2f-032f27957617)
This is the vision for the more advanced level where the students learn
to iterate/sort the array.