Blockly: refactor 'shootFireball' (BlocklyCommands) (#2005)
This is a follow-up to #1909 (0116e47), were methods like
`shootFireball()` and `aimAndShoot()` got moved from `Server` to
`BlocklyCommands`.

Unfortunately, some attributes are no longer available in the new
auxiliary class `BlocklyCommands`, so the old (moved) code now has to
fetch some things multiple times.

This PR performs a refactoring of the `shootFireball` method in
`BlocklyCommands` to increase efficiency and readbility of this method.