Blockly: refactor BlocklyCommands#active() (#2121)
This is a follow-up to #2110, where changes to the `active()` method in
[BlocklyCommands](url) were introduced. This required multiple
terminating and reopening operations on the stream, which is not
particularly attractive. In addition, the value `null` was also returned
in the auxiliary method `active()`, which forces all callers to use the
usual scheme `var wuppie = active(); if (wuppie == null) foo();`.

This refactoring introduces an `Optional<Tile>` as return of `active()`,
which should lead to a better readability and maintainability of the
code.

Additionally, some convenience methods were defined to make the code
more readable and increase the symmetry of the operations.

---

~~edit: waiting for #2120 (`Direction` gets moved)~~

---------

Co-authored-by: Niklas Schumann <68606032+Flamtky@users.noreply.github.com>