Refactoring: clean-up '.toCoordinate().toCenteredPoint()' (#2233)
The class `core.utils.Point` now has an auxiliary method `Point
toCenteredPoint()`. There is no longer any reason to first translate the
point into a coordinate and then calculate the centre point on this
coordinate. This can now be done directly on the `Point` itself.

This PR replaces calls of `.toCoordinate().toCenteredPoint()` on objects
of type `Point` with `.toCenteredPoint()`.

fixes #2231