LevelUtils: refactor randomAccessibleTileCoordinateInRange to improve readability (#2238)
The method `LevelUtils#randomAccessibleTileCoordinateInRange` was
returning the `Coordinate` of a random tile in range. However, all uses
would convert the result to `Point` - so let's do this right in the
method itself.

Also, just use `findAny()` instead of a custom randomisation. According
to the
[doc](https://docs.oracle.com/en/java/javase/21/docs/api/java.base/java/util/stream/Stream.html#findAny())
`findAny()` is "_explicitly nondeterministic_", thus it will return a
random element from the stream.