Refactor DebugDrawSystem
Instead of using the Camera batch and scaling the font to draw in world coordinates, it now converts the world coordinates to screen coordinates and renders in screen space. This allows fonts to not look like trash, since they remain unscaled. Also added generic drawText methods which can be used by other systems to quickly draw text on screen / in world.